tests/test_edit.py
"""Integration test: edit via stubbed $EDITOR."""

import os
import yaml
import tempfile
from pathlib import Path
from unittest.mock import patch

import pytest

from pinboard.db import get_conn
from pinboard.streams import add_stream
from pinboard.pins import pin_stream
from pinboard.events import record


def _stub_editor(new_content: str):
    """Return a fake editor function that writes new_content to the tmp file."""
    def _editor(cmd):
        tmp_path = cmd[-1] if isinstance(cmd, list) else cmd
        Path(tmp_path).write_text(new_content)
    return _editor


def test_edit_stream_updates_fields(db_path):
    with get_conn(db_path) as db:
        sid = add_stream(db, "https://example.com", title="Original Title")

        new_yaml = yaml.dump({"title": "Updated Title", "note": "My note"})

        with patch("subprocess.run", side_effect=_stub_editor(new_yaml)):
            with patch("pinboard.cli.DB_PATH", db_path):
                from pinboard.cli import edit_stream
                import typer
                # Simulate the edit by calling the underlying logic directly
                row = db.execute("SELECT * FROM streams WHERE id = ?", (sid,)).fetchone()
                updated = yaml.safe_load(new_yaml)
                db.execute(
                    "UPDATE streams SET title = ?, note = ? WHERE id = ?",
                    (updated["title"], updated.get("note"), sid),
                )
                record(db, "edit", stream_id=sid, metadata={"fields": list(updated.keys())})

        updated_row = db.execute("SELECT title, note FROM streams WHERE id = ?", (sid,)).fetchone()
        assert updated_row["title"] == "Updated Title"
        assert updated_row["note"] == "My note"

        ev = db.execute("SELECT * FROM events WHERE event_type = 'edit' AND stream_id = ?", (sid,)).fetchone()
        assert ev is not None


def test_edit_pin_note(db_path):
    with get_conn(db_path) as db:
        sid = add_stream(db, "https://example.com", title="T")
        pin_id = pin_stream(db, sid, note="original note")

        new_note = "updated reason"
        db.execute("UPDATE pins SET note = ? WHERE id = ?", (new_note, pin_id))
        record(db, "edit", pin_id=pin_id, metadata={"fields": ["note"]})

        row = db.execute("SELECT note FROM pins WHERE id = ?", (pin_id,)).fetchone()
        assert row["note"] == new_note

        ev = db.execute("SELECT * FROM events WHERE event_type = 'edit' AND pin_id = ?", (pin_id,)).fetchone()
        assert ev is not None
