pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pinboard"
version = "0.1.0"
description = "A local-first personal pinning & streams system"
requires-python = ">=3.9"
dependencies = [
    "typer>=0.9",
    "rich>=13",
    "anthropic>=0.25",
    "openai>=1.0",
    "trafilatura>=1.7",
    "pypdf>=3.0",
    "numpy>=1.24",
    "tomli>=2.0; python_version < '3.11'",
    "pyyaml>=6.0",
    "httpx>=0.23",
]

[project.optional-dependencies]
local-embeddings = ["sentence-transformers>=3.0"]
dev = ["pytest>=8", "pytest-mock>=3.12", "ruff>=0.4"]

[project.scripts]
pinboard = "pinboard.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
